Try adding xmaps and omaps for next and previous commands
The xmaps don't really work. Going to the next/previous TODO works, but
extending the selection doesn't because of the way I set it up with the
mark system. We'd probably need a more involved function here to do what
we want.

Currently doesn't support counts. I'll have to add that in.