Add new motions to move between regular and important TODOs
The existing `[u` and `]u` bindings allow moving to the previous and
next unfinished to-do, but when there are many low priority to-dos, and
a handful of regular and high priority items between them, those
commands are ineffective for quickly moving to the relevant ones.

Add two new sets of commands:

* `[o` & `]o` to move between regular priority to-dos
* `[y` & `]y` to move between high priority to-dos

I was looking for bracket mappings that aren't bound by default. Using
`o` seems to work, meaning "ordinary". And `y` could refer to
"yimportant". Close enough.