doc/todo.txt: Replace backtick separators with quotes
Backticks are used to refer to Vim commands in help files. The Vim docs
explain:

> When referring to a Vim command and to create a hot-link, place the
> name between two backticks, eg. inside `:filetype`.  You will see this
> is highlighted as a command, like a code block (see below).

These aren't Vim commands, so use a different surrounding mark. Settled
on double quotes.

This also fixes a display problem with the "`- `" text, which was
rendered as "`- " (without the final backtick) in the highlighted help
due to the parsing for syntax highlight concealing.